+++ libs/dsp/src/Buffers.cpp
@@ -1,9 +1,10 @@
 #include "dsp/Buffers.h"
 
+#include <cstddef>
 #include <cstdint>
 
 namespace dsp::buffers {
-void initStereoBuffer(StereoBuffer& buffer, uint32_t size) {
+void initStereoBuffer(StereoBuffer& buffer, size_t size) {
   buffer.buffer = new float[size * 2]();
   buffer.left = buffer.buffer;
   buffer.right = buffer.left + size;
